src/requirements.txt
django-filter
Django==3.0.2
django-crispy-forms==1.8.1
django-widget-tweaks==1.4.5
Pillow==7.0.0
django-tables2==2.2.1
django-bootstrap4==1.1.1